moved implementation to CodeBuilderFinalizerImpl
